#60 Prepare to merge dev to main
Adding stub analytics endpoint
